Revert "gallium: make handles of set_global_binding 64 bit"
This reverts commit e1ffb72a05f9b50ee47767aaadbab3e47896ee14

+++ src/gallium/include/pipe/p_context.h
@@ -804,7 +804,7 @@ struct pipe_context {
    void (*set_global_binding)(struct pipe_context *context,
                               unsigned first, unsigned count,
                               struct pipe_resource **resources,
-                              uint64_t **handles);
+                              uint32_t **handles);
 
    /**
     * Launch the compute kernel starting from instruction \a pc of the
